--- src/main/java/site/iotify/userservice/advice/UserControllerAdvice.java
@@ -1,40 +1,0 @@
-package site.iotify.userservice.advice;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import site.iotify.userservice.exception.UnAuthenticatedException;
-import site.iotify.userservice.exception.UnAuthorizedException;
-import site.iotify.userservice.exception.UserNotFoundException;
-
-@ControllerAdvice
-@Slf4j
-public class UserControllerAdvice {
-
-    @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleUserNotFoundException(UserNotFoundException e) {
-        log.error(e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-    }
-
-    @ExceptionHandler(UnAuthorizedException.class)
-    public ResponseEntity<ErrorResponse> handleUnAuthorizedException(UnAuthorizedException e) {
-        log.error(e.getMessage());
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
-
-    @ExceptionHandler(UnAuthenticatedException.class)
-    public ResponseEntity<ErrorResponse> handleUnAuthenticatedException(UnAuthenticatedException e) {
-        log.error(e.getMessage());
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        log.error(e.getMessage());
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
-    }
-}
