+++ src/main/java/site/iotify/userservice/domain/user/dto/ChangePasswordRequest.java
@@ -1,0 +1,20 @@
+package site.iotify.userservice.domain.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+
+@Getter
+public class ChangePasswordRequest {
+
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String oldPassword;
+
+    @NotBlank
+    private String newPassword;
+
+    @NotBlank
+    private String confirmPassword;
+}
