+++ src/main/java/site/iotify/userservice/domain/user/dto/ChirpstackTenantUserDto.java
@@ -1,0 +1,11 @@
+package site.iotify.userservice.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ChirpstackTenantUserDto {
+    private ChirpstackUserInfo tenantUser;
+    private String tenantId;
+}
