+++ src/main/java/site/iotify/userservice/domain/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package site.iotify.userservice.domain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import site.iotify.userservice.domain.user.entity.User;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, String> {
+
+    Optional<User> findByEmail(String email);
+}
