--- src/main/java/site/iotify/userservice/dto/ChangePasswordRequest.java
@@ -1,20 +1,0 @@
-package site.iotify.userservice.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-@Getter
-public class ChangePasswordRequest {
-
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String oldPassword;
-
-    @NotBlank
-    private String newPassword;
-
-    @NotBlank
-    private String confirmPassword;
-}
