--- src/main/java/site/iotify/userservice/dto/ChirpstackTenantUserDto.java
@@ -1,11 +1,0 @@
-package site.iotify.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ChirpstackTenantUserDto {
-    private ChirpstackUserInfo tenantUser;
-    private String tenantId;
-}
