--- src/main/java/site/iotify/userservice/exception/CreationFailedException.java
@@ -1,4 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class CreationFailedException extends RuntimeException {
-}
