--- src/main/java/site/iotify/userservice/exception/TenantNotFoundException.java
@@ -1,7 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class TenantNotFoundException extends RuntimeException {
-    public TenantNotFoundException(String s) {
-        super(s);
-    }
-}
