--- src/main/java/site/iotify/userservice/exception/UnAuthenticatedException.java
@@ -1,4 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class UnAuthenticatedException extends RuntimeException {
-}
