--- src/main/java/site/iotify/userservice/exception/UnAuthorizedException.java
@@ -1,4 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class UnAuthorizedException extends RuntimeException {
-}
