--- src/main/java/site/iotify/userservice/exception/UserAlreadyExistsException.java
@@ -1,8 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class UserAlreadyExistsException extends RuntimeException {
-
-    public UserAlreadyExistsException(String userId) {
-        super("User with id " + userId + " already exists");
-    }
-}
