--- src/main/java/site/iotify/userservice/exception/UserNotFoundException.java
@@ -1,12 +1,0 @@
-package site.iotify.userservice.exception;
-
-public class UserNotFoundException extends RuntimeException {
-
-    public UserNotFoundException(Long id) {
-        super("User not found with id " + id);
-    }
-
-    public UserNotFoundException(String email) {
-        super("Could not find user with email " + email);
-    }
-}
