+++ src/main/java/site/iotify/userservice/global/exception/CreationFailedException.java
@@ -1,0 +1,4 @@
+package site.iotify.userservice.global.exception;
+
+public class CreationFailedException extends RuntimeException {
+}
