+++ src/main/java/site/iotify/userservice/global/exception/TenantNotFoundException.java
@@ -1,0 +1,7 @@
+package site.iotify.userservice.global.exception;
+
+public class TenantNotFoundException extends RuntimeException {
+    public TenantNotFoundException(String s) {
+        super(s);
+    }
+}
