+++ src/main/java/site/iotify/userservice/global/exception/UnAuthorizedException.java
@@ -1,0 +1,4 @@
+package site.iotify.userservice.global.exception;
+
+public class UnAuthorizedException extends RuntimeException {
+}
