+++ src/main/java/site/iotify/userservice/global/exception/UserAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package site.iotify.userservice.global.exception;
+
+public class UserAlreadyExistsException extends RuntimeException {
+
+    public UserAlreadyExistsException(String userId) {
+        super("User with id " + userId + " already exists");
+    }
+}
