--- src/main/java/site/iotify/userservice/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package site.iotify.userservice.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import site.iotify.userservice.entity.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, String> {
-
-    Optional<User> findByEmail(String email);
-}
